+++ runtime/litentry/src/tests/mod.rs
@@ -15,5 +15,5 @@
 // along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
 
 mod base_call_filter;
-// mod setup;
+
 mod transaction_payment;
