--- runtime/litentry/src/tests/setup.rs
@@ -1,104 +1,0 @@
-// Copyright 2020-2022 Litentry Technologies GmbH.
-// This file is part of Litentry.
-//
-// Litentry is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Litentry is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
-
-use frame_support::weights::{DispatchInfo, PostDispatchInfo, Weight};
-pub use pallet_balances::Call as BalancesCall;
-
-pub use crate::{
-	AccountId, Balance, Balances, Call, Event, ExtrinsicFilter, Multisig, Origin, ParachainSystem,
-	Runtime, System, TransactionByteFee, TransactionPayment, Treasury, Vesting,
-};
-
-pub const ALICE: [u8; 32] = [1u8; 32];
-pub const BOB: [u8; 32] = [2u8; 32];
-pub const CHARLIE: [u8; 32] = [3u8; 32];
-
-pub(crate) fn alice() -> AccountId {
-	AccountId::from(ALICE)
-}
-
-pub(crate) fn bob() -> AccountId {
-	AccountId::from(BOB)
-}
-
-pub(crate) fn charlie() -> AccountId {
-	AccountId::from(CHARLIE)
-}
-
-pub struct ExtBuilder {
-	balances: Vec<(AccountId, Balance)>,
-	parachain_id: u32,
-}
-
-impl Default for ExtBuilder {
-	fn default() -> Self {
-		Self { balances: vec![], parachain_id: 2008 }
-	}
-}
-
-impl ExtBuilder {
-	pub fn balances(mut self, balances: Vec<(AccountId, Balance)>) -> Self {
-		self.balances = balances;
-		self
-	}
-
-	#[allow(dead_code)]
-	pub fn parachain_id(mut self, parachain_id: u32) -> Self {
-		self.parachain_id = parachain_id;
-		self
-	}
-
-	pub fn build(self) -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::default().build_storage::<Runtime>().unwrap();
-		pallet_balances::GenesisConfig::<Runtime> {
-			balances: self
-				.balances
-				.into_iter()
-				.map(|(account_id, initial_balance)| (account_id, initial_balance))
-				.collect::<Vec<_>>(),
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		let mut ext = sp_io::TestExternalities::new(t);
-		ext.execute_with(|| System::set_block_number(1));
-		ext
-	}
-}
-
-/// create a transaction info struct from weight. Handy to avoid building the whole struct.
-pub fn info_from_weight(w: Weight) -> DispatchInfo {
-	// pays_fee: Pays::Yes -- class: DispatchClass::Normal
-	DispatchInfo { weight: w, ..Default::default() }
-}
-
-pub fn post_info_from_weight(w: Weight) -> PostDispatchInfo {
-	PostDispatchInfo { actual_weight: Some(w), pays_fee: Default::default() }
-}
-
-pub fn run_with_system_weight<F>(w: Weight, mut assertions: F)
-where
-	F: FnMut(),
-{
-	let mut t: sp_io::TestExternalities = frame_system::GenesisConfig::default()
-		.build_storage::<Runtime>()
-		.unwrap()
-		.into();
-	t.execute_with(|| {
-		System::set_block_consumed_resources(w, 0);
-		assertions()
-	});
-}
